SmallDataIO: Combine reader and writer
Both read and write functions are now performed with a single
class, SmallDataIO
