+++ Source/IO/SmallDataIO.cpp
@@ -7,6 +7,7 @@
 #include "SmallDataIO.hpp"
 
 #include <AMReX_ParallelDescriptor.H>
+#include <AMReX_Print.H>
 #include <AMReX_Utility.H>
 #include <AMReX_Vector.H>
 
@@ -16,8 +17,6 @@
 // instead of cstdio but I'm sure someone would tell me off for not maintaining
 // backwards compatability.
 #include <cstdio> // for std::rename and std::remove
-#include <iomanip>
-#include <iostream>
 #include <sstream>
 #include <utility>
 
@@ -29,7 +28,9 @@ const std::string SmallDataIO::s_default_file_extension = ".dat";
 
 SmallDataIO::SmallDataIO(const std::string &a_filename_prefix, double a_dt,
                          double a_time, double a_restart_time, Mode a_mode,
-                         bool a_first_step, const std::string &a_file_extension,
+                         bool a_first_step,
+                         const file_structure_t *a_file_structure,
+                         const std::string &a_file_extension,
                          int a_data_precision, int a_coords_precision,
                          int a_filename_steps_width)
     : m_filename(a_filename_prefix + a_file_extension), m_dt(a_dt),
@@ -42,10 +43,11 @@ SmallDataIO::SmallDataIO(const std::string &a_filename_prefix, double a_dt,
       m_coords_precision(a_coords_precision),
       m_coords_width(m_coords_precision + 5),
       m_coords_epsilon(std::pow(10.0, -a_coords_precision)),
-      m_rank(amrex::ParallelDescriptor::MyProc())
+      // this is for reading only
+      m_structure_defined{false}
 {
 
-    if (m_rank == 0)
+    if (amrex::ParallelDescriptor::IOProcessor())
     {
         std::ios::openmode file_openmode = std::ios::out;
         if (m_mode == APPEND)
@@ -94,11 +96,33 @@ SmallDataIO::SmallDataIO(const std::string &a_filename_prefix, double a_dt,
             amrex::UtilRenameDirectoryToOld(m_filename, call_mpi_barrier);
         }
         m_file.open(m_filename, file_openmode);
+
         if (!m_file)
         {
-            amrex::Abort("SmallDataIO::error opening file for writing");
+            amrex::Abort("SmallDataIO::error opening file " + m_filename);
+        }
+        if (m_mode == READ)
+        {
+            read_file();
+            if (a_file_structure == nullptr)
+                determine_file_structure();
+            else
+            {
+                set_file_structure(*a_file_structure);
+            }
         }
     }
+}
+
+SmallDataIO::SmallDataIO(const std::string &a_filename_prefix, double a_dt,
+                         double a_time, double a_restart_time, Mode a_mode,
+                         bool a_first_step, const std::string &a_file_extension,
+                         int a_data_precision, int a_coords_precision,
+                         int a_filename_steps_width)
+    : SmallDataIO(a_filename_prefix, a_dt, a_time, a_restart_time, a_mode,
+                  a_first_step, nullptr, a_file_extension, a_data_precision,
+                  a_coords_precision, a_filename_steps_width)
+{
 }
 
 SmallDataIO::SmallDataIO(const std::string &a_filename_prefix, double a_dt,
@@ -107,7 +131,7 @@ SmallDataIO::SmallDataIO(const std::string &a_filename_prefix, double a_dt,
                          int a_data_precision, int a_coords_precision,
                          int a_filename_steps_width)
     : SmallDataIO(a_filename_prefix, a_dt, a_time, a_restart_time, a_mode,
-                  (a_time == a_dt), a_file_extension, a_data_precision,
+                  (a_time == a_dt), nullptr, a_file_extension, a_data_precision,
                   a_coords_precision, a_filename_steps_width)
 {
 }
@@ -115,19 +139,36 @@ SmallDataIO::SmallDataIO(const std::string &a_filename_prefix, double a_dt,
 SmallDataIO::SmallDataIO(const std::string &a_filename_prefix,
                          const std::string &a_file_extension,
                          int a_data_precision, int a_coords_precision)
-    : SmallDataIO(a_filename_prefix, 0.0, 0.0, 0.0, READ, false,
+    : SmallDataIO(a_filename_prefix, 0.0, 0.0, 0.0, READ, false, nullptr,
                   a_file_extension, a_data_precision, a_coords_precision, 0)
+{
+}
+
+SmallDataIO::SmallDataIO(const std::string &a_filename_prefix,
+                         const file_structure_t *a_file_structure,
+                         const std::string &a_file_extension,
+                         int a_data_precision, int a_coords_precision)
+    : SmallDataIO(a_filename_prefix, 0.0, 0.0, 0.0, READ, false,
+                  a_file_structure, a_file_extension, a_data_precision,
+                  a_coords_precision, 0)
 {
 }
 
 //! Destructor (closes file)
 SmallDataIO::~SmallDataIO()
 {
-    if (m_rank == 0)
+    m_filename.clear();
+    m_structure_defined = false;
+
+    if (amrex::ParallelDescriptor::IOProcessor())
     {
         m_file.close();
+        m_file_structure.clear();
+        m_file_contents.clear();
     }
 }
+
+// NOLINTEND(bugprone-easily-swappable-parameters)
 
 // ------------ Writing Functions ------------
 
@@ -143,9 +184,11 @@ void SmallDataIO::write_header_line(
     write_header_line(a_header_strings, pre_header_strings);
 }
 
+// NOLINTBEGIN(bugprone-easily-swappable-parameters)
 void SmallDataIO::write_header_line(
     const std::vector<std::string> &a_header_strings,
     const std::vector<std::string> &a_pre_header_strings)
+// NOLINTEND(bugprone-easily-swappable-parameters)
 {
     if (m_rank == 0)
     {
@@ -185,8 +228,10 @@ void SmallDataIO::write_time_data_line(const std::vector<double> &a_data)
     write_data_line(a_data, m_time);
 }
 
+// NOLINTBEGIN(bugprone-easily-swappable-parameters)
 void SmallDataIO::write_data_line(const std::vector<double> &a_data,
                                   const std::vector<double> &a_coords)
+// NOLINTEND(bugprone-easily-swappable-parameters)
 {
     if (m_rank == 0)
     {
@@ -254,6 +299,433 @@ void SmallDataIO::remove_duplicate_time_data(const bool keep_m_time_data)
 
 // ------------ Reading Functions ------------
 
+void SmallDataIO::file_structure_t::clear()
+{
+    num_blocks = 0;
+    block_starts.clear();
+    num_data_rows.clear();
+    num_header_rows.clear();
+    num_data_columns.clear();
+}
+
+// Only for use in the constructor - this assumes that the file is already open
+void SmallDataIO::read_file()
+{
+
+    AMREX_ASSERT_WITH_MESSAGE(
+        m_file, "This function can only be called from the constructor! \n");
+
+    m_file.seekg(0, std::ios::end);
+
+    auto size = m_file.tellg();
+
+    AMREX_ALWAYS_ASSERT_WITH_MESSAGE(
+        size < max_file_size,
+        "SmallDataIO: File size is too large. Max file size is 1 GB. \n");
+
+    std::string file_contents(size,
+                              '\0'); // construct string to stream size
+    m_file.seekg(0);
+    m_file.read(file_contents.data(), size);
+    AMREX_ALWAYS_ASSERT_WITH_MESSAGE(!file_contents.empty(),
+                                     "SmallDataIO: File contents of " +
+                                         m_filename + " could not be read");
+
+    m_file_contents = file_contents;
+}
+
+// Parses the file and determines its structure
+// Don't call this directly, it will automatically be called when you open a
+// file
+void SmallDataIO::determine_file_structure()
+{
+    // first check that the file has been read
+    AMREX_ASSERT(m_file_contents.empty() == false);
+
+    std::istringstream file_stream(m_file_contents);
+    // go through each line and determine structure
+    std::string line;
+    int block_counter        = 0; // assume we always have one block
+    auto current_position    = file_stream.tellg();
+    int block_start_position = current_position;
+    int header_row_counter   = 0;
+    int data_row_counter     = 0;
+
+    while (std::getline(file_stream, line))
+    {
+        if (!line.empty())
+        {
+            // header rows start with '#'
+            if (line.find('#') == 0)
+            {
+                if (header_row_counter == 0)
+                {
+                    block_start_position = current_position;
+                }
+
+                ++header_row_counter;
+            }
+            else
+            {
+                if (data_row_counter++ == 0)
+                {
+
+                    // only count a new block if it contains a data row
+                    m_file_structure.block_starts.push_back(
+                        block_start_position);
+                    ++block_counter;
+                    // determine column structure from first data row in
+                    // block get a vector of the widths of the columns
+                    // including preceeding whitespace
+                    std::string::size_type start_whitespace = 0;
+
+                    int ncols{0};
+                    while (!(start_whitespace == std::string::npos))
+                    {
+                        std::string::size_type start_non_whitespace =
+                            line.find_first_not_of(' ', start_whitespace);
+                        std::string::size_type next_start_whitespace =
+                            line.find_first_of(' ', start_non_whitespace);
+                        start_whitespace = next_start_whitespace;
+                        ncols++;
+                    }
+
+                    m_file_structure.num_data_columns.push_back(ncols);
+                }
+            }
+        }
+        else
+        {
+            if (header_row_counter > 0 || data_row_counter > 0)
+
+            {
+                // end of previous block
+                m_file_structure.num_header_rows.push_back(header_row_counter);
+                m_file_structure.num_data_rows.push_back(data_row_counter);
+                // reset the counters
+                header_row_counter = 0;
+                data_row_counter   = 0;
+            }
+        }
+        current_position = file_stream.tellg();
+    }
+    // Just in case the file ends without a line break:
+    if (data_row_counter > 0)
+    {
+        m_file_structure.num_header_rows.push_back(header_row_counter);
+        m_file_structure.num_data_rows.push_back(data_row_counter);
+    }
+
+    m_file_structure.num_blocks = block_counter;
+
+    AMREX_ASSERT(m_file_structure.num_data_rows.size() ==
+                 m_file_structure.num_blocks);
+    AMREX_ASSERT(m_file_structure.num_header_rows.size() ==
+                 m_file_structure.num_blocks);
+    AMREX_ASSERT(m_file_structure.num_data_columns.size() ==
+                 m_file_structure.num_blocks);
+    AMREX_ASSERT(m_file_structure.block_starts.size() ==
+                 m_file_structure.num_blocks);
+
+    m_structure_defined = true;
+}
+
+// Set structure if known already (e.g. same as another file already
+// determined)
+void SmallDataIO::set_file_structure(
+    const SmallDataIO::file_structure_t &a_file_structure)
+{
+    m_file_structure    = a_file_structure;
+    m_structure_defined = true;
+}
+
+// File struture getter
+const SmallDataIO::file_structure_t &SmallDataIO::get_file_structure() const
+{
+    return m_file_structure;
+}
+
+// Utility for viewing the file struture
+
+void SmallDataIO::print_file_structure() const
+{
+    if (amrex::ParallelDescriptor::IOProcessor())
+    {
+        AMREX_ASSERT(m_structure_defined);
+
+        amrex::Print() << "Total number of blocks: "
+                       << m_file_structure.num_blocks << "\n";
+
+        for (int i = 0; i < m_file_structure.num_blocks; i++)
+        {
+
+            // Printing info on file structure:
+            amrex::Print() << "#######" << "\n";
+            amrex::Print() << "Info on Block " << i << ":" << "\n";
+            amrex::Print() << "Block start: "
+                           << m_file_structure.block_starts[i] << "\n";
+            amrex::Print() << "Number of columns in Block " << i << ": "
+                           << m_file_structure.num_data_columns[i] << "\n";
+            amrex::Print() << "Number of rows in Block " << i << ": "
+                           << m_file_structure.num_data_rows[i] << "\n";
+            amrex::Print() << "Number of header rows in Block " << i << ": "
+                           << m_file_structure.num_header_rows[i] << "\n";
+
+            // Printing column names:
+
+            std::vector<std::string> header;
+            get_header_strings(header, i);
+
+            amrex::Print() << "Header strings found: ";
+            for (auto &column_name : header)
+            {
+                amrex::Print() << column_name << " ";
+            }
+            amrex::Print() << "\n";
+            amrex::Print() << "#######" << "\n";
+        }
+    }
+}
+
+// Get an interval of columns (inclusive) from a block
+// NOLINTBEGIN(bugprone-easily-swappable-parameters)
+void SmallDataIO::get_columns(std::vector<SmallDataIO::column_t> &out,
+                              int a_min_column, int a_max_column, int a_block)
+// NOLINTEND(bugprone-easily-swappable-parameters)
+{
+    if (amrex::ParallelDescriptor::IOProcessor())
+    {
+        AMREX_ASSERT(m_file_contents.empty() == false);
+        AMREX_ASSERT(m_structure_defined);
+        AMREX_ASSERT(0 <= a_min_column);
+        AMREX_ASSERT(a_max_column <=
+                     m_file_structure.num_data_columns[a_block]);
+        AMREX_ASSERT(a_min_column <= a_max_column);
+        const int num_columns = a_max_column - a_min_column + 1;
+
+        out.resize(num_columns);
+        for (auto &column : out)
+        {
+            column.resize(m_file_structure.num_data_rows[a_block]);
+        }
+
+        std::istringstream file_stream(m_file_contents);
+        skip_ahead(file_stream, m_file_structure.num_header_rows[a_block],
+                   a_block);
+
+        double discard = 0.0;
+        for (int irow = 0; irow < m_file_structure.num_data_rows[a_block];
+             ++irow)
+        {
+
+            for (int icolumn = 0;
+                 icolumn < m_file_structure.num_data_columns[a_block];
+                 ++icolumn)
+            {
+                if (a_min_column <= icolumn && icolumn <= a_max_column)
+                {
+
+                    file_stream >> out[icolumn - a_min_column][irow];
+                }
+                else
+                {
+                    file_stream >> discard;
+                }
+            }
+        }
+    }
+}
+
+// Get columns from a block based on their names in the header
+// NOLINTBEGIN(bugprone-easily-swappable-parameters)
+void SmallDataIO::get_columns(std::vector<SmallDataIO::column_t> &out,
+                              const std::vector<std::string> &column_names,
+                              const int a_block)
+// NOLINTEND(bugprone-easily-swappable-parameters)
+{
+
+    if (amrex::ParallelDescriptor::IOProcessor())
+    {
+        // Resize for the number of requested columns
+        out.resize(column_names.size());
+
+        // Resize for the total number of rows
+        for (auto &column : out)
+        {
+            column.resize(m_file_structure.num_data_rows[a_block]);
+        }
+
+        AMREX_ASSERT(m_file_contents.empty() == false);
+        AMREX_ASSERT(m_structure_defined);
+
+        // Get column names from header
+        // Assume that the column names are in the line immediately proceeding
+        // the data, like this:
+        // # blah blah this is my data blah blah
+        // # blah blah more things to say blah blah
+        // # x y z
+
+        std::vector<std::string> header;
+        get_header_strings(header, a_block);
+
+        // Associate each column name with the number
+        // of where they are in the file
+
+        std::vector<int> column_numbers;
+
+        for (const auto &column_name : column_names)
+        {
+            int column_no = 0;
+            for (const auto &header_name : header)
+            {
+                if (header_name == column_name)
+                {
+
+                    column_numbers.push_back(column_no);
+                    break;
+                }
+                column_no++;
+            }
+
+            // If all header_names have been checked but column_name has not
+            // been found...
+            if (column_no == m_file_structure.num_data_columns[a_block])
+            {
+                std::string error_message{
+                    column_name + " could not be read. Please double check "
+                                  "your inputs to SmallDataIO"};
+                amrex::Abort(error_message);
+            }
+        }
+
+        std::istringstream file_stream(m_file_contents);
+
+        skip_ahead(file_stream, m_file_structure.num_header_rows[a_block],
+                   a_block);
+
+        double discard = 0.0;
+        for (int irow = 0; irow < m_file_structure.num_data_rows[a_block];
+             ++irow)
+        {
+            // This is a bit slow but returns the columns in the order that the
+            // user requested them
+
+            for (int icolumn = 0;
+                 icolumn < m_file_structure.num_data_columns[a_block];
+                 ++icolumn)
+            {
+                auto const column_ptr = std::find(
+                    column_numbers.begin(), column_numbers.end(), icolumn);
+                if (column_ptr != column_numbers.end())
+                {
+                    auto column_index = column_ptr - column_numbers.begin();
+                    file_stream >> out[column_index][irow];
+                }
+                else
+                {
+                    file_stream >> discard;
+                }
+            }
+        }
+    }
+}
+
+// Get a data column from a block
+void SmallDataIO::get_all_data_columns(std::vector<SmallDataIO::column_t> &out,
+                                       int a_block)
+{
+    assert(m_structure_defined);
+    int min_data_column = 0;
+    int max_data_column = m_file_structure.num_data_columns[a_block];
+    get_columns(out, min_data_column, max_data_column, a_block);
+}
+
+void SmallDataIO::get_column(std::vector<SmallDataIO::column_t> &out,
+                             int a_column, int a_block)
+{
+    get_columns(out, a_column, a_column, a_block);
+}
+
+// Returns a vector of numeric values from a header row
+// NOLINTBEGIN(bugprone-easily-swappable-parameters)
+void SmallDataIO::get_data_from_header(std::vector<amrex::Real> &out,
+                                       int a_header_row_number,
+                                       int a_block) const
+// NOLINTEND(bugprone-easily-swappable-parameters)
+{
+    if (amrex::ParallelDescriptor::IOProcessor())
+    {
+
+        assert(m_file_contents.empty() == false);
+        assert(m_structure_defined);
+        assert(a_header_row_number < m_file_structure.num_header_rows[a_block]);
+
+        std::istringstream file_stream(m_file_contents);
+        skip_ahead(file_stream, a_header_row_number, a_block);
+
+        // get desired header line
+        std::string line;
+        std::getline(file_stream, line);
+
+        // find numbers in header using regex
+        // I think this takes a long time to compile...
+        std::regex number("[+-]?([0-9]*\\.)?[0-9]+");
+        auto numbers_begin =
+            std::sregex_iterator(line.begin(), line.end(), number);
+        auto numbers_end = std::sregex_iterator();
+
+        for (std::sregex_iterator rit = numbers_begin; rit != numbers_end;
+             ++rit)
+        {
+            // put matches in vector
+            out.push_back(std::stod((*rit).str()));
+        }
+    }
+}
+
+// NOLINTBEGIN(bugprone-easily-swappable-parameters)
+void SmallDataIO::get_header_strings(std::vector<std::string> &header,
+                                     int a_block) const
+// NOLINTEND(bugprone-easily-swappable-parameters)
+{
+
+    if (amrex::ParallelDescriptor::IOProcessor())
+    {
+        assert(m_file_contents.empty() == false);
+        assert(m_structure_defined);
+
+        header.resize(m_file_structure.num_data_columns[0]);
+
+        std::istringstream file_stream(m_file_contents);
+        int nlines_to_skip = m_file_structure.num_header_rows[a_block] - 1;
+        skip_ahead(file_stream, nlines_to_skip, a_block);
+
+        std::string hashes; // remove the #s at the beginning of the line
+        file_stream >> hashes;
+
+        for (auto &icol : header)
+        {
+            file_stream >> icol;
+        }
+    }
+}
+// Helper function to skip lines e.g. header lines
+// NOLINTBEGIN(bugprone-easily-swappable-parameters)
+void SmallDataIO::skip_ahead(std::istringstream &file_stream,
+                             int nlines_to_skip, int a_block) const
+// NOLINTEND(bugprone-easily-swappable-parameters)
+{
+    // rewind to the beginning of the block
+    file_stream.seekg(m_file_structure.block_starts[a_block], std::ios::beg);
+
+    // assume header rows are at start of block
+    for (int irow = 0; irow < nlines_to_skip; ++irow)
+    {
+        // skip lines before desired row
+        file_stream.ignore(std::numeric_limits<std::streamsize>::max(), '\n');
+    }
+}
+
 void SmallDataIO::get_specific_data_line(std::vector<double> &a_out_data,
                                          const std::vector<double> &a_coords)
 {
@@ -314,10 +786,12 @@ void SmallDataIO::get_specific_data_line(std::vector<double> &a_out_data,
 
 // ------------ Other Functions --------------
 
+// NOLINTBEGIN(bugprone-easily-swappable-parameters)
 std::string SmallDataIO::get_new_filename(const std::string &a_file_prefix,
                                           double a_dt, double a_time,
                                           const std::string &a_file_extension,
                                           int a_filename_steps_width)
+// NOLINTEND(bugprone-easily-swappable-parameters)
 {
     AMREX_ASSERT(a_dt > 0);
     const int step = static_cast<int>(std::round(a_time / a_dt));
@@ -355,4 +829,30 @@ double SmallDataIO::get_default_coords_epsilon()
     return pow(10.0, -s_default_coords_precision);
 }
 
-// NOLINTEND(bugprone-easily-swappable-parameters)
+// Helper function to redistribute data amongst all ranks
+void SmallDataIO::broadcast_data(std::vector<SmallDataIO::column_t> &data)
+{
+    int nrows{0};
+    int ncols{0};
+
+    if (amrex::ParallelDescriptor::IOProcessor())
+    {
+        nrows = data[0].size();
+        ncols = data.size();
+    }
+    amrex::ParallelDescriptor::Bcast(
+        &ncols, 1, amrex::ParallelDescriptor::IOProcessorNumber());
+
+    amrex::ParallelDescriptor::Bcast(
+        &nrows, 1, amrex::ParallelDescriptor::IOProcessorNumber());
+
+    data.resize(ncols);
+
+    for (auto &column : data)
+    {
+        column.resize(nrows);
+        amrex::ParallelDescriptor::Bcast(
+            column.data(), nrows,
+            amrex::ParallelDescriptor::IOProcessorNumber());
+    }
+}
