+++ Tests/SmallDataIOTest/SmallDataIOTest.cpp
@@ -38,18 +38,16 @@ bool check_almost_equal(std::vector<double> vector_1,
 
     for (int i = 0; i < vector_1.size(); ++i)
     {
-        if (std::abs(vector_1[i] - vector_2[i]) > err_tol)
-        {
-            return false;
-        }
+        CHECK_MESSAGE(vector_1[i] ==
+                          doctest::Approx(vector_2[i]).epsilon(err_tol),
+                      "i= " << i);
     }
 
     return true;
 }
 // NOLINTBEGIN(bugprone-easily-swappable-parameters)
-void test_small_data_io_writer(
-    const std::vector<SmallDataIOReader::column_t> &col,
-    const int data_precision)
+void test_small_data_io_writer(const std::vector<SmallDataIO::column_t> &col,
+                               const int data_precision)
 // NOLINTEND(bugprone-easily-swappable-parameters)
 {
     const std::string filename_prefix{"test_"};
@@ -73,12 +71,10 @@ void test_small_data_io_writer(
     }
 }
 
-std::vector<SmallDataIOReader::column_t>
+std::vector<SmallDataIO::column_t>
 test_small_data_io_reader(const std::vector<std::string> &column_names)
 {
-
-    SmallDataIOReader test_reader;
-    test_reader.open("test_000000.dat");
+    SmallDataIO test_reader("test_000000");
 
     // Could print out file structure as well
     // test_reader.print_file_structure();
@@ -93,26 +89,29 @@ test_small_data_io_reader(const std::vector<std::string> &column_names)
     // If a column name doesn't exist in the header, amrex::Abort will
     // be called
 
-    std::vector<SmallDataIOReader::column_t> data;
+    std::vector<SmallDataIO::column_t> data;
     test_reader.get_columns(data, column_names, 0);
 
-    SmallDataIOReader::broadcast_data(data);
+    SmallDataIO::broadcast_data(data);
 
     return data;
 }
 
 // NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
-std::vector<SmallDataIOReader::column_t>
+std::vector<SmallDataIO::column_t>
 test_small_data_io_reader(const int a_min_col, const int a_max_col)
 {
+    // If we know the data structure already we can pass it in like so:
 
-    SmallDataIOReader test_reader;
-    test_reader.open("test_000000.dat");
+    const SmallDataIO::file_structure_t test_file_structure{
+        1, {0}, {1}, {100}, {4}};
 
-    std::vector<SmallDataIOReader::column_t> data;
+    SmallDataIO test_reader("test_000000", &test_file_structure);
+
+    std::vector<SmallDataIO::column_t> data;
     test_reader.get_columns(data, a_min_col, a_max_col, 0);
 
-    SmallDataIOReader::broadcast_data(data);
+    SmallDataIO::broadcast_data(data);
 
     return data;
 }
@@ -120,10 +119,9 @@ test_small_data_io_reader(const int a_min_col, const int a_max_col)
 std::vector<double> test_small_data_io_reader(const int a_col)
 
 {
-    SmallDataIOReader test_reader;
-    test_reader.open("test_000000.dat");
+    SmallDataIO test_reader("test_000000");
 
-    std::vector<SmallDataIOReader::column_t> data;
+    std::vector<SmallDataIO::column_t> data;
     test_reader.get_column(data, a_col, 0);
 
     return data[0];
@@ -147,7 +145,7 @@ void run_small_data_io_test()
         amrex::InitRandom(cpu_seed, amrex::ParallelDescriptor::NProcs(),
                           gpu_seed);
 
-        std::vector<SmallDataIOReader::column_t> write_data(3);
+        std::vector<SmallDataIO::column_t> write_data(3);
         for (auto &column : write_data)
         {
             column.resize(Npts);
@@ -159,8 +157,7 @@ void run_small_data_io_test()
         static constexpr int data_precision = 10;
         static const amrex::Real err_tol = std::pow(10., -1.0 * data_precision);
 
-        // Test the file writer: write out the random co-oridinates
-
+        // Test the file writer: write out the random coordinates
         test_small_data_io_writer(write_data, data_precision);
 
         // Test the file reader: read the random numbers back in using their
@@ -168,17 +165,16 @@ void run_small_data_io_test()
         std::vector<std::string> read_these_columns{"x", "z"};
         auto read_data1 = test_small_data_io_reader(read_these_columns);
 
-        // Test the file reader - read the data by specifying the column numbers
+        // Test the file reader: read the data by specifying the column numbers
         const int min_col = 1;
         const int max_col = 3;
         auto read_data2   = test_small_data_io_reader(min_col, max_col);
 
-        // Test if this part satisfies linter complexity rule
-        CHECK(check_almost_equal(read_data1[0], write_data[0], err_tol));
-        CHECK(check_almost_equal(read_data1[1], write_data[2], err_tol));
-        CHECK(check_almost_equal(read_data2[0], write_data[0], err_tol));
-        CHECK(check_almost_equal(read_data2[1], write_data[1], err_tol));
-        CHECK(check_almost_equal(read_data2[2], write_data[2], err_tol));
+        check_almost_equal(read_data1[0], write_data[0], err_tol);
+        check_almost_equal(read_data1[1], write_data[2], err_tol);
+        check_almost_equal(read_data2[0], write_data[0], err_tol);
+        check_almost_equal(read_data2[1], write_data[1], err_tol);
+        check_almost_equal(read_data2[2], write_data[2], err_tol);
     }
 
     amrex::Finalize();
