+++ Tests/SmallDataIOTest/SmallDataIOTest.hpp
@@ -10,7 +10,6 @@
 
 // GRTeclyn includes
 #include "SmallDataIO.hpp"
-#include "SmallDataIOReader.hpp"
 
 // AMReX includes
 #include <AMReX.H>
@@ -31,16 +30,15 @@ bool check_almost_equal(std::vector<double> vector_1,
 std::vector<double> generate_random_numbers(const int Npts);
 
 // NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
-void test_small_data_io_writer(
-    const std::vector<SmallDataIOReader::column_t> &col,
-    const int data_precision);
+void test_small_data_io_writer(const std::vector<SmallDataIO::column_t> &col,
+                               const int data_precision);
 
 // NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
-std::vector<SmallDataIOReader::column_t>
+std::vector<SmallDataIO::column_t>
 test_small_data_io_reader(const std::vector<std::string> &column_names);
 
 // NOLINTNEXTLINE(bugprone-easily-swappable-parameters)
-std::vector<SmallDataIOReader::column_t>
+std::vector<SmallDataIO::column_t>
 test_small_data_io_reader(const int a_min_col, const int a_max_col);
 
 std::vector<double> test_small_data_io_reader(const int a_col);
